requirements.txt
qrcode[pil]
Pillow
Flask
